src/azure-ai-102/information-extraction/04-knowledge-mining/pyproject.toml
[project]
name = "04-knowledge-mining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-search-documents==11.6.0",
    "python-dotenv>=1.2.2",
]
